+++ R/compat.R
@@ -36,3 +36,28 @@ deparse <- function(expr, width.cutoff = 500, ...) {
 readChar <- function (con, nchars, useBytes = FALSE) {
   suppressWarnings(base::readChar(con, nchars, useBytes))
 }
+
+# to avoid incompatibilities between newer versions of queryparser and older
+# versions of tidyquery, call this function at the top of every queryparser
+# function that is imported by tidyquery
+assert_tidyquery_version <- function(min_version = "0.2.0") {
+  # is the function that called this function being called from tidyquery?
+  if (identical(get0(".packageName", topenv(parent.frame(2)), inherits = FALSE), "tidyquery")) {
+    current_version <-
+      mget(
+        ".tidyquery.version",
+        envir = asNamespace("tidyquery"),
+        ifnotfound = list(package_version("0.0.0")),
+        inherits = FALSE
+      )[[1]]
+    # if yes, error if tidyquery version too old
+    if (current_version < package_version(min_version)) {
+      stop(
+        "Incompatible tidyquery version. Install tidyquery ",
+        min_version,
+        " or higher",
+        call. = FALSE
+      )
+    }
+  }
+}